Extract AddCustomChainModal to its own component
