+++ packages/nextjs/components/NetworksDropdown/index.ts
@@ -1,2 +1,3 @@
+export * from "./AddCustomChainModal";
 export * from "./CustomOption";
 export * from "./OtherChainsModal";
